CYCU-ChatBot/requirements.txt
# Flask Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# LangChain and AI
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
openai==1.6.1

# Vector Store and Embeddings
faiss-cpu==1.7.4
# faiss-gpu==1.7.4  # 如果有 GPU，可改用此版本

# PDF Processing
pypdf==3.17.4
PyPDF2==3.0.1

# Environment Variables
python-dotenv==1.0.0

# Ngrok (for development)
pyngrok==7.0.5

# HTTP Requests
requests==2.31.0

# Data Processing
numpy==1.24.3
pandas==2.0.3

# Utilities
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Code Quality
black==23.12.1
flake8==7.0.0
pylint==3.0.3

# Production Server
gunicorn==21.2.0
waitress==2.1.2

# Logging
colorlog==6.8.0

# Type Checking (optional)
mypy==1.8.0
